Move the symbol following code into a separate class ExportAnalyzer.ts
