+++ apps/api-extractor/src/analyzer/SymbolAnalyzer.ts
@@ -5,8 +5,6 @@
 
 import * as ts from 'typescript';
 
-import { TypeScriptHelpers } from './TypeScriptHelpers';
-
 /**
  * This is a helper class for DtsRollupGenerator and AstSymbolTable.
  * Its main role is to provide an expanded version of TypeScriptHelpers.followAliases()
@@ -47,34 +45,6 @@ export class SymbolAnalyzer {
     }
 
     return false;
-  }
-
-  public static isAmbient(symbol: ts.Symbol, typeChecker: ts.TypeChecker): boolean {
-    const followedSymbol: ts.Symbol = TypeScriptHelpers.followAliases(symbol, typeChecker);
-
-    if (followedSymbol.declarations && followedSymbol.declarations.length > 0) {
-      const firstDeclaration: ts.Declaration = followedSymbol.declarations[0];
-
-      // Test 1: Are we inside the sinister "declare global {" construct?
-      const highestModuleDeclaration: ts.ModuleDeclaration | undefined
-        = TypeScriptHelpers.findHighestParent(firstDeclaration, ts.SyntaxKind.ModuleDeclaration);
-      if (highestModuleDeclaration) {
-        if (highestModuleDeclaration.name.getText().trim() === 'global') {
-          return true;
-        }
-      }
-
-      // Test 2: Otherwise, the main heuristic for ambient declarations is by looking at the
-      // ts.SyntaxKind.SourceFile node to see whether it has a symbol or not (i.e. whether it
-      // is acting as a module or not).
-      const sourceFileNode: ts.Node | undefined = TypeScriptHelpers.findFirstParent(
-        firstDeclaration, ts.SyntaxKind.SourceFile);
-      if (sourceFileNode && !!typeChecker.getSymbolAtLocation(sourceFileNode)) {
-        return false;
-      }
-    }
-
-    return true;
   }
 
   /*
